separated out id to id manager, exclude use of static attributes
