+++ src/main/java/entities/IDManager.java
@@ -8,52 +8,39 @@ public class IDManager {
     /**
      * The current maximum ID of all the users.
      */
-    private static int userId;
+    private int userId;
 
     /**
      * The current maximum ID of all the studies.
      */
-    private static int studyId;
+    private int studyId;
 
     /**
      * The current maximum ID of all the questionnaires.
      */
-    private static int questionnaireId;
+    private int questionnaireId;
 
     /**
      * The current maximum ID of all the questions.
      */
-    private static int questionId;
+    private int questionId;
 
     /**
      * The current maximum ID of all the answers.
      */
-    private static int answerId;
+    private int answerId;
 
     /**
      * The current maximum ID of all the versioned answers.
      */
-    private static int versionedAnswerId;
+    private int versionedAnswerId;
 
-
-    /**
-     * Constructor for IDManager.
-     * @param initialValue  the initial value of all the IDs
-     */
-    public IDManager(int initialValue) {
-        userId = initialValue;
-        studyId = initialValue;
-        questionnaireId = initialValue;
-        questionId = initialValue;
-        answerId = initialValue;
-        versionedAnswerId = initialValue;
-    }
 
     /**
      * Appends the current maximum ID of all the users by 1 and returns the new ID.
      * @return  the new ID
      */
-    public static int newUserId() {
+    public int newUserId() {
         return ++userId;
     }
 
@@ -62,7 +49,7 @@ public class IDManager {
      * Appends the current maximum ID of all the studies by 1 and returns the new ID.
      * @return  the new ID
      */
-    public static int newStudyId() {
+    public int newStudyId() {
         return ++studyId;
     }
 
@@ -71,7 +58,7 @@ public class IDManager {
      * Appends the current maximum ID of all the questionnaires by 1 and returns the new ID.
      * @return  the new ID
      */
-    public static int newQuestionnaireId() {
+    public int newQuestionnaireId() {
         return ++questionnaireId;
     }
 
@@ -80,7 +67,7 @@ public class IDManager {
      * Appends the current maximum ID of all the questions by 1 and returns the new ID.
      * @return  the new ID
      */
-    public static int newQuestionId() {
+    public int newQuestionId() {
         return ++questionId;
     }
 
@@ -89,7 +76,7 @@ public class IDManager {
      * Appends the current maximum ID of all the answers by 1 and returns the new ID.
      * @return  the new ID
      */
-    public static int newAnswerId() {
+    public int newAnswerId() {
         return ++answerId;
     }
 
@@ -98,7 +85,7 @@ public class IDManager {
      * Appends the current maximum ID of all the versioned answers by 1 and returns the new ID.
      * @return  the new ID
      */
-    public static int newVersionedAnswerId() {
+    public int newVersionedAnswerId() {
         return ++versionedAnswerId;
     }
 }
