+++ src/main/java/entities/MultipleChoiceQuestion.java
@@ -28,8 +28,8 @@ public class MultipleChoiceQuestion extends Question implements Serializable {
      * @param variableName      Keyword that describes the content of this Question
      * @param content           What is being asked the participants.
      */
-    public MultipleChoiceQuestion(Questionnaire questionnaire, String variableName, String content) {
-        super(questionnaire, variableName, content);
+    public MultipleChoiceQuestion(int id, Questionnaire questionnaire, String variableName, String content) {
+        super(id, questionnaire, variableName, content);
     }
 
 
@@ -53,72 +53,6 @@ public class MultipleChoiceQuestion extends Question implements Serializable {
         this.numOfChoices++;
         this.choices.add(choice);
         return true;
-    }
-
-
-    /**
-     * Removes a choice from the list of choices for this MultipleChoiceQuestion.
-     * This method is overloaded.
-     *
-     * @param choice    The choice to be removed.
-     * @return true if choice was successfully removed from the list of choices.
-     */
-    public boolean removeChoice(String choice) {
-        if (this.choices.contains(choice)) {
-            this.choices.remove(choice);
-            numOfChoices--;
-            return true;
-        }
-        return false;
-    }
-
-
-    /**
-     * Removes the choice at index [index] from the list of choices for this MultipleChoiceQuestion.
-     * This method is overloaded.
-     *
-     * @param index     The index of the choice to be removed.
-     * @return true if choice at index was successfully removed from the list of choices.
-     */
-    public boolean removeChoice(int index) {
-        if (this.isInChoices(index)) {
-            this.choices.remove(index);
-            numOfChoices--;
-            return true;
-        }
-        return false;
-    }
-
-    /**
-     * Rearrange the list of choices so that the choice at index [indexToArranged] is inserted in [index].
-     *
-     * @param indexToArrange    The index of the choice to be rearranged.
-     * @param index             The index to insert the choice at.
-     * @return true if choice at index indexToArranged was successfully inserted at index.
-     */
-    public boolean rearrangeChoices(int indexToArrange, int index) {
-        if (this.isInChoices(indexToArrange) && this.isInChoices(index)) {
-            String holdChoice = this.choices.remove(indexToArrange);
-            this.choices.add(index, holdChoice);
-            return true;
-        }
-        return false;
-    }
-
-
-    /**
-     * Modify the choice in index so that the newChoice is in this index.
-     *
-     * @param index     The index of the choice to be modified.
-     * @param newChoice The new choice to be inserted.
-     * @return true if the choice at index was successfully modified.
-     */
-    public boolean modifyChoice(int index, String newChoice) {
-        if (this.isInChoices(index)) {
-            this.choices.set(index, newChoice);
-            return true;
-        }
-        return false;
     }
