+++ src/main/java/entities/QuestionFactory.java
@@ -3,12 +3,10 @@ package entities;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.io.Serializable;
-
 /**
  * A QuestionFactory that creates and returns Question objects.
  */
-public class QuestionFactory implements QuestionFactoryInterface, Serializable {
+public class QuestionFactory {
 
     /**
      * The create method that gets called with a specified type of Question when creating a Question object.
@@ -20,14 +18,14 @@ public class QuestionFactory implements QuestionFactoryInterface, Serializable {
      * @param content           The content of the Question.
      * @return                  The Question object that was created.
      */
-    public static @Nullable Question create(@NotNull String type, Questionnaire questionnaire, String variableName, String content) {
+    public static @Nullable Question create(int id, @NotNull String type, Questionnaire questionnaire, String variableName, String content) {
         switch (type) {
             case "MC":
-                return new MultipleChoiceQuestion(questionnaire, variableName, content);
+                return new MultipleChoiceQuestion(id, questionnaire, variableName, content);
             case "Scale":
-                return new ScaleQuestion(questionnaire, variableName, content);
+                return new ScaleQuestion(id, questionnaire, variableName, content);
             case "Text":
-                return new TextQuestion(questionnaire, variableName, content);
+                return new TextQuestion(id, questionnaire, variableName, content);
         }
         return null;
     }
