+++ src/main/java/entities/User.java
@@ -27,10 +27,10 @@ public class User implements Serializable {
      * @param username the username of this User
      * @param name the name of this User
      */
-    public User(String username, String name) {
+    public User(int id, String username, String name) {
         this.username = username;
         this.name = name;
-        this.id = IDManager.newUserId();
+        this.id = id;
     }
 
     /**
