+++ src/main/java/entities/UserFactory.java
@@ -18,11 +18,11 @@ public class UserFactory implements UserFactoryInterface, Serializable {
      * @param name      of the user
      * @return the new User object
      */
-    public User create(@NotNull String type, String username, String name){
+    public User create(int id, @NotNull String type, String username, String name){
         switch (type){
-            case "Researcher" : return new Researcher(username, name);
-            case "Participant" : return new Participant(username, name);
+            case "Researcher" : return new Researcher(id, username, name);
+            case "Participant" : return new Participant(id, username, name);
+            default: return null;
         }
-        return null;
     }
 }
