+++ src/main/java/use_cases/answer_questionnaire/AnswerQuestionnaireInteractor.java
@@ -10,7 +10,15 @@ import java.util.Objects;
 
 public class AnswerQuestionnaireInteractor implements AnswerQuestionnaireInputBoundary {
 
+    /**
+     * The presenter which this interactor uses to output the response.
+     */
     private AnswerQuestionnaireOutputBoundary answerQuestionnairePresenter;
+
+    /**
+     * The ID manager which this interactor uses to fetch the IDs of the entities.
+     */
+    private IDManager idManager;
 
 
     /**
@@ -89,8 +97,10 @@ public class AnswerQuestionnaireInteractor implements AnswerQuestionnaireInputBo
         } else if (! checkAnswers(questionnaire, answers, numQuestions)) {
             answerQuestionnairePresenter.presentAnswerQuestionnaireFailure("The answers were not valid.");
         } else {
-            Answer newAnswer = new Answer(participant, questionnaire);
-            VersionedAnswer newVersionedAnswer = new VersionedAnswer(1, user, answers, newAnswer);
+            int newAnswerId = idManager.newAnswerId();
+            Answer newAnswer = new Answer(newAnswerId, participant, questionnaire);
+            int newVersionedAnswerId = idManager.newVersionedAnswerId();
+            VersionedAnswer newVersionedAnswer = new VersionedAnswer(newVersionedAnswerId, 1, user, answers, newAnswer);
             newAnswer.addNewVersion(newVersionedAnswer);
             return newAnswer;
         }
@@ -124,6 +134,15 @@ public class AnswerQuestionnaireInteractor implements AnswerQuestionnaireInputBo
      */
     public void setAnswerQuestionnairePresenter(AnswerQuestionnaireOutputBoundary answerQuestionnairePresenter) {
         this.answerQuestionnairePresenter = answerQuestionnairePresenter;
+    }
+
+
+    /**
+     * Set the ID manager for this interactor.
+     * @param idManager The ID manager.
+     */
+    public void setIdManager(IDManager idManager) {
+        this.idManager = idManager;
     }
