+++ src/main/java/use_cases/create_questionnaire/CreateQuestion.java
@@ -4,7 +4,6 @@ import entities.*;
 import org.jetbrains.annotations.NotNull;
 import user_interface_layer.screens.create_questionnaire_inputs_screen.QuestionModel;
 
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -16,10 +15,11 @@ public class CreateQuestion {
      * @param questionnaire The questionnaire to add the questions to.
      * @param questionData The data for the questions to be createdd and added.
      */
-    public static void createQuestion(Questionnaire questionnaire, @NotNull QuestionModel questionData) {
+    public static void createQuestion(int newId, Questionnaire questionnaire, @NotNull QuestionModel questionData) {
         String type = questionData.getType();
         Question question =
-                QuestionFactoryInterface.create(
+                QuestionFactory.create(
+                        newId,
                         type,
                         questionnaire,
                         questionData.getVariable(),
@@ -29,8 +29,6 @@ public class CreateQuestion {
             for (String singleOption : options) {
                 assert question != null;
                 ((MultipleChoiceQuestion) question).addChoice(singleOption);
-                System.out.println(singleOption);
-
             }
 
         } else if (type.equals("Scale")) {
@@ -39,8 +37,6 @@ public class CreateQuestion {
             ((ScaleQuestion) question).setBottomLabel(options.get(0));
             ((ScaleQuestion) question).setTopLabel(options.get(1));
             ((ScaleQuestion) question).modifyScaleRange(Integer.parseInt(options.get(2)));
-            System.out.println(options.get(2));
-
         }
         assert question != null;
         questionnaire.addQuestion(question);
